+++ frontend/src/openapi/v2/models/AuthorizationDB.ts
@@ -8,9 +8,6 @@ import type { RoleType } from './RoleType';
  * The creator of the Authorization object should also be the creator of the dataset itself.
  */
 export type AuthorizationDB = {
-    creator: string;
-    created?: string;
-    modified?: string;
     dataset_id: string;
     user_ids?: Array<string>;
     role: RoleType;
