+++ frontend/src/openapi/v2/models/EventListenerIn.ts
@@ -6,7 +6,7 @@
  * On submission, minimum info for a listener is name, version and description. Clowder will use name and version to locate queue.
  */
 export type EventListenerIn = {
-    author?: string;
+    creator?: string;
     name: string;
     version?: string;
     description?: string;
