+++ frontend/src/openapi/v2/models/FeedOut.ts
@@ -4,17 +4,26 @@
 
 import type { FeedListener } from './FeedListener';
 import type { SearchObject } from './SearchObject';
-import type { UserOut } from './UserOut';
 
 /**
- * A Job Feed is a saved set of search criteria with some number of Event Listeners that can be triggered when new
- * resources match the saved search criteria for the Feed.
+ * Document Mapping class.
+ *
+ * Fields:
+ *
+ * - `id` - MongoDB document ObjectID "_id" field.
+ * Mapped to the PydanticObjectId class
+ *
+ * Inherited from:
+ *
+ * - Pydantic BaseModel
+ * - [UpdateMethods](https://roman-right.github.io/beanie/api/interfaces/#aggregatemethods)
  */
 export type FeedOut = {
-    id?: string;
+    creator: string;
+    created?: string;
+    modified?: string;
     name: string;
     search: SearchObject;
     listeners?: Array<FeedListener>;
-    author?: UserOut;
-    updated?: string;
+    _id?: string;
 }
