+++ frontend/src/openapi/v2/models/MetadataConfig.ts
@@ -3,6 +3,5 @@
 /* eslint-disable */
 
 export type MetadataConfig = {
-    id?: string;
     type?: string;
 }
