+++ frontend/src/openapi/v2/models/MetadataEnumConfig.ts
@@ -3,7 +3,6 @@
 /* eslint-disable */
 
 export type MetadataEnumConfig = {
-    id?: string;
     type?: string;
     options: Array<string>;
 }
