+++ frontend/src/openapi/v2/models/MetadataField.ts
@@ -6,7 +6,6 @@ import type { MetadataConfig } from './MetadataConfig';
 import type { MetadataEnumConfig } from './MetadataEnumConfig';
 
 export type MetadataField = {
-    id?: string;
     name: string;
     list?: boolean;
     widgetType?: string;
