+++ frontend/src/openapi/v2/models/MetadataIn.ts
@@ -7,7 +7,6 @@ import type { ExtractorInfo } from './ExtractorInfo';
 import type { LegacyEventListenerIn } from './LegacyEventListenerIn';
 
 export type MetadataIn = {
-    id?: string;
     context?: Array<string>;
     context_url?: string;
     definition?: string;
