+++ frontend/src/openapi/v2/models/MetadataOut.ts
@@ -19,12 +19,12 @@ import type { MongoDBRef } from './MongoDBRef';
  * - [UpdateMethods](https://roman-right.github.io/beanie/api/interfaces/#aggregatemethods)
  */
 export type MetadataOut = {
-    _id?: string;
     context?: Array<string>;
     context_url?: string;
     definition?: string;
     content: any;
     description?: string;
+    _id?: string;
     resource: MongoDBRef;
     agent: MetadataAgent;
     created?: string;
