+++ frontend/src/openapi/v2/models/Repository.ts
@@ -6,7 +6,6 @@
  * Reference to a repository associated with Event Listener/Extractor.
  */
 export type Repository = {
-    id?: string;
     repository_type?: string;
     repository_url?: string;
 }
